Made multiple changes in support of push notifications
I'm using two different libraries for push notifications. One that
wraps the base ReactNative functionality for iOS and one that
implements Firebase Cloud Messaging support for Android. Notifications
are currently received while the app is open or closed. If the app is
open and in the foreground then a local notifation has to be made
and pushed to the device, if that is desired.