Figured out unit testing with jest and redux.
Jest is a nice testing framework, mocking objects is built in. I was
able to ignore the parts of a function that I should not be testing,
such as responses from TRACS. Those are mocked with the various possible
responses to test how my logic responds to them. Also, Promises suck.