Added settings, settings saving works. Settings recall does not.
I may have to rethink the use of a class in this. ES6 classes can be modified outside the world of redux. If I use a class as one of the state objects then I can change it without calling an action which seems to be in direct violation of redux, and is failing to trigger certain component lifecycle events.