Lined up home icon reasonably well. Removed other icons after header is collapsed
I removed the extremely excessive calling of a redux action every time the user scrolled.
Once an animated value is in the store it is still bound to the scroll event. The reference
to it only needs to go in the store once and the header will track scrolling until the
component is unmounted.