Resolved a multitude of release build issues on both platforms
Android: For some reason the release bundle was being created but not
copied to the assets folder. The release build had a bundle from 9/27
instead of the most recent one. Also matched Android version to 1.2.3
and updated internal version to 12.

iOS: There were (as always) some linking errors related to third party
dependencies, and the version number had to be updated. Info.plist is
also using <string></string> tags instead of <string/> when the field
is empty.