Added AX state for comments
The comments are "hidden" off screen but the screen reader will still
swipe over to them so they had to be explicitly shown when the user
selects view comments. The grade item was also still visible to the
screen reader when the comments were shown so the AX state is tracked
for both.