Fixed gradebook item spacing. Fixed logout bug.
Gradebook items were spaced too close together. Increased the margin
to 5px from 3px. Had to adjust the comment absolute positioning as well.

The menu option match on logging out has probably been broken for a
while but I rarely log out. The conditional was matching on lower case
'logout' and the value is actually "Logout" so the match was always
false.